+++ src/components/AddressBar/index.tsx
@@ -4,17 +4,10 @@ import { useHistory } from 'react-router'
 import styled from 'styled-components'
 import { ReactComponent as QrcodeSvg } from '../../assets/svg/qrcode.svg'
 import { RoutePath } from '../../routes'
-import { copyFallback, truncateMiddle } from '../../utils'
+import { copyFallback, generateOldAddress, truncateMiddle } from '../../utils'
 import { useToast } from '../../hooks/useToast'
 import { trackLabels, useTrackClick } from '../../hooks/useTrack'
-import { useAccount, WalletType } from '../../hooks/useAccount'
-import {
-  AddressPrefix,
-  addressToScript,
-  fullPayloadToAddress,
-  AddressType,
-} from '@nervosnetwork/ckb-sdk-utils'
-import { IS_MAINNET } from '../../constants'
+import { useAccount } from '../../hooks/useAccount'
 
 const Container = styled.div`
   height: 32px;
@@ -62,19 +55,8 @@ export const Addressbar: React.FC<AddressbarProps> = ({
   const trackCopy = useTrackClick('home', 'click')
   const { walletType } = useAccount()
   const displayAddress = useMemo(() => {
-    if (walletType === WalletType.Flashsigner) {
-      return address
-    }
-    const script = addressToScript(address)
-    return fullPayloadToAddress({
-      args: script.args,
-      type:
-        script.hashType === 'data'
-          ? AddressType.DataCodeHash
-          : AddressType.TypeCodeHash,
-      codeHash: script.codeHash,
-      prefix: IS_MAINNET ? AddressPrefix.Mainnet : AddressPrefix.Testnet,
-    })
+    // eslint-disable-next-line @typescript-eslint/no-non-null-assertion
+    return generateOldAddress(address, walletType!)
   }, [walletType, address])
   return (
     <Container className="address-bar">
